requirements-txt.txt
pandas==2.0.0
numpy==1.24.3
matplotlib==3.7.1
requests==2.31.0
python-telegram-bot==20.4
